backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
opencv-python==4.9.0.80
numpy==1.26.4
python-multipart==0.0.6
Pillow==10.4.0